Feature/csharp app

+++ csharp_winforms/pcb_monitoring_program/pcb_monitoring_program/DatabaseManager/Models/DefectDetail.cs
@@ -2,15 +2,17 @@ using System;
 
 namespace pcb_monitoring_program.DatabaseManager.Models
 {
-    /// 상세 불량 유형별 검출 정보 모델 (defect_details 테이블)
-    /// YOLO가 검출한 상세 클래스별 통계 정보 저장
+    /// <summary>
+    /// 상세 불량 유형별 검출 정보 모델 (v3.1 스키마)
+    /// YOLO가 검출한 상세 클래스별 통계 정보 저장 (트리거로 자동 삽입)
+    /// </summary>
     public class DefectDetail
     {
         /// 고유 ID
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         /// 검사 ID (외래 키)
-        public int InspectionId { get; set; }
+        public long InspectionId { get; set; }
 
         /// YOLO 검출 클래스명 (예: solder_bridge, capacitor_missing, poor_solder 등)
         public string ClassName { get; set; }
